make it a class
